requirements.txt
ijson
mysql-connector